# api.yaml
swagger: "2.0"
info:
  title: Setly API
  description: This is Setly API Documentation.
  version: 1.0.0
  contact:
    email: essiensaviour.a@gmail.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/liscences/LIS"
  servers:
    - url: http://localhost:4000/api/v1
      description: Local server URL
    - url: https://api.setly.ai/api/v1
      description: Live server URL
schemes: [https,http]
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags:
        - Auth
      summary: Signup a new user
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
           $ref: "#/components/schemas/Register"
      responses:
        201:
          description: Successful Response
        400:
          description: User Error
        500:
          description: Internal Server Error




#Authenticate User
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      description: An existing user with account.
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          schema:
           $ref: "#/components/schemas/Auth"
      responses:
        200:
          description: Successful Response
        400:
          description: User Error
        500:
          description: Internal Server Error


#Verify User Email
...
  /list:
    get:
      tags:
        - Prospect List  
      summary: Fetch prospect list
      description: Already authenticated user
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:    # status code
          description: Successul Response
        400:
          description: User Error
        500:
          description: Internal Server Error


#Create a Prospect List
...
  /prospects/:
    put:
      tags:
        - Prospect
      summary: Update a single prospect
      description: Already authenticated user
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/components/schemas/Prospect"
      responses:
        200:    # status code
          description: Successul Response
        400:
          description: User Error
        500:
          description: Internal Server Error



#Fetch a prospect
...
  /prospects:
    get:
      tags:
        - Prospect  
      summary: List prospects
      description: Already authenticated user
      produces:
        - application/json
      consumes:
        - application/json
      responses:
        200:    # status code
          description: Successul Response
        400:
          description: User Error
        500:
          description: Internal Server Error













#Start a campaign
...
components:
  schemas:
    Register:
      type: object
      required:
        - first_name
        - last_name
        - email
        - password
        - company
        - industry
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        password:
          type: string
        company_name:
          type: string
        industry:
          type: string
    Auth:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
...
    Prospect:
      type: object
      required:
        - id
        - name
        - page
        - first_name
        - last_name
        - url
        - handle
      properties:
        id:
          type: string
        name:
          type: string
        page:
          type: number
        first_name:
          type: string
        last_name:
          type: string
        url:
          type: string
        email:
          type: string
        phone:
          type: string
        region:
          type: string
        job_title:
          type: string
        photo:
          type: string
        handle:
          type: string
        headline:
          type: string
        company:
          type: string
        connection:
          type: string
        website:
          type: array
...